Move InAppMessaging into a sub-category under Notifications
