+++ packages/aws-amplify/src/index.ts
@@ -44,7 +44,7 @@ export {
 export { PubSub } from '@aws-amplify/pubsub';
 export { default as Cache } from '@aws-amplify/cache';
 export { Interactions } from '@aws-amplify/interactions';
-export { InAppMessaging } from '@aws-amplify/notifications';
+export { Notifications } from '@aws-amplify/notifications';
 export * from '@aws-amplify/ui';
 export { XR } from '@aws-amplify/xr';
 export { Predictions } from '@aws-amplify/predictions';
