+++ packages/notifications/src/inAppMessaging/index.ts
@@ -1,0 +1,3 @@
+export { default } from './InAppMessaging';
+export * from './types';
+export * from './Providers';
