+++ packages/notifications/src/index.ts
@@ -1,3 +1,2 @@
-export { default as InAppMessaging } from './InAppMessaging';
+export { default as Notifications } from './Notifications';
 export * from './types';
-export * from './Providers';
