+++ packages/notifications/src/types.ts
@@ -10,94 +10,11 @@
  * CONDITIONS OF ANY KIND, either express or implied. See the License for the specific language governing permissions
  * and limitations under the License.
  */
-export type OnMessagesReceived = (messages: any) => void;
+import { InAppMessagingConfig } from './inAppMessaging';
+export * from './inAppMessaging/types';
 
-export type InAppMessagingCategory = 'InAppMessaging';
+export type NotificationsCategory = 'Notifications';
 
-export type InAppMessagingEvent = {
-	name: string;
-	attributes?: Record<string, string>;
-	metrics?: Record<string, number>;
-};
-
-export interface InAppMessagingConfig {
-	listenForAnalyticsEvents?: boolean;
-	onMessagesReceived?: OnMessagesReceived;
-}
-
-export interface InAppMessagingProvider {
-	// you need to implement these methods
-
-	// configure your provider
-	configure(config: object): object;
-
-	// return 'InAppMessaging';
-	getCategory(): InAppMessagingCategory;
-
-	// return the name of you provider
-	getProviderName(): string;
-
-	// get in-app messages from provider
-	getInAppMessages(): Promise<any>;
-
-	// filters in-app messages based on event input and provider logic
-	processInAppMessages(
-		messages: InAppMessage[],
-		event: InAppMessagingEvent
-	): Promise<InAppMessage[]>;
-}
-
-export type InAppMessageLayout =
-	| 'BOTTOM_BANNER'
-	| 'CAROUSEL'
-	| 'MIDDLE_BANNER'
-	| 'OVERLAYS'
-	| 'TOP_BANNER';
-
-export type InAppMessageAction = 'CLOSE' | 'DEEP_LINK' | 'LINK';
-
-interface InAppMessageHeader {
-	content: string;
-	style?: InAppMessageStyle;
-}
-
-interface InAppMessageBody {
-	content: string;
-	style?: InAppMessageStyle;
-}
-
-interface InAppMessageImage {
-	src: string;
-}
-
-interface InAppMessageButton {
-	title: string;
-	action: InAppMessageAction;
-	url?: string;
-	style?: InAppMessageStyle;
-}
-
-export interface InAppMessageStyle {
-	backgroundColor?: string;
-	borderRadius?: number;
-	color?: string;
-	textAlign?: 'center' | 'left' | 'right';
-}
-
-export interface InAppMessageContent {
-	header?: InAppMessageHeader;
-	body?: InAppMessageBody;
-	image?: InAppMessageImage;
-	primaryButton?: InAppMessageButton;
-	secondaryButton?: InAppMessageButton;
-}
-
-export interface InAppMessage {
-	id: string;
-	layout: InAppMessageLayout;
-	content: InAppMessageContent[];
-	metadata?: any;
-	onAction?: () => void;
-	onDismiss?: () => void;
-	onDisplay?: () => void;
+export interface NotificationsConfig {
+	InAppMessaging?: InAppMessagingConfig;
 }
